menial fixes
