+++ client/src/pages/CreateCampaign.tsx
@@ -41,8 +41,7 @@ export default function () {
         })
     }
 
-    const handleSubmit: MouseEventHandler<HTMLButtonElement> = function (e) {
-        e.preventDefault();
+    const handleSubmit: MouseEventHandler<HTMLButtonElement> = function () {
         setIsLoading(true);
 
         try {
