+++ client/src/pages/CreateMission.tsx
@@ -28,8 +28,7 @@ export default function () {
         })
     }
 
-    const handleSubmit: MouseEventHandler<HTMLButtonElement> = function (e) {
-        e.preventDefault();
+    const handleSubmit: MouseEventHandler<HTMLButtonElement> = function () {
         setIsLoading(true);
 
         try {
