+++ client/src/pages/Login.tsx
@@ -36,8 +36,7 @@ export default function (): ReactElement {
         })
     }
 
-    const handleSubmit: MouseEventHandler<HTMLButtonElement> = function (e) {
-        e.preventDefault();
+    const handleSubmit: MouseEventHandler<HTMLButtonElement> = function () {
         try {
             validateData<UserLogin>(loginData)
             dispatch(login(loginData))
